requirements.txt
chardet==2.3.0
Django==1.8.4
django-contrib-comments==1.7.3
django-crispy-forms==1.6.0
django-filter==0.15.3
django-mptt==0.8.6
django-registration-redux==1.4
enum==0.4.6
grizzled-python==1.0.1
MySQL-python==1.2.3
ndg-httpsclient==0.4.2
Pillow==3.3.1
prettytable==0.7.2
pyasn1==0.1.9
pyOpenSSL==16.1.0
python-weixin==0.1.6.1
PyYAML==3.12
requests==2.11.1
sae-python-dev==1.3.6
simplejson==3.10.0
six==1.10.0
django-breadcrumbs==1.1.3
django-ckeditor==5.2.2
django-pagination==1.0.7
olefile==0.44
packaging==16.8
pyparsing==2.2.0
sqlcmd==0.7.1
wechat-sdk==0.5.8
wechatpay==1.1.5
Werkzeug==0.11.11


appdirs==1.4.3
Django==1.8
django-breadcrumbs==1.1.3
django-ckeditor==5.2.2
django-pagination==1.0.7
olefile==0.44
packaging==16.8
Pillow==4.1.0
pyparsing==2.2.0
six==1.10.0
